Add LovosKortele component
